backend/pyproject.toml
[project]
name = "spotify-analytics-backend"
version = "0.1.0"
description = "Spotifyプレイリスト分析ツール - バックエンド"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.3",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "spotipy>=2.23.0",
    "pandas>=2.1.4",
    "numpy>=1.26.3",
    "scikit-learn>=1.4.1",
    "sqlalchemy>=2.0.23",
    "celery>=5.3.4",
    "redis>=5.0.1",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
